[pacer] CSV feed now derive from ipacer
